+++ src/components/StructuredContent/StructuredContentTypes.tsx
@@ -1,0 +1,5 @@
+export enum StructuredContentImagePosition {
+  Left = "left",
+  Right = "right",
+  Center = "center",
+}