+++ src/theme/index.ts
@@ -27,6 +27,7 @@ import Icon from "./components/icon";
 import Label from "./components/label";
 import Link from "./components/link";
 import List from "./components/list";
+import Logo from "./components/logo";
 import NotificationStyles from "./components/notification";
 import Pagination from "./components/pagination";
 import ProgressIndicator from "./components/progressIndicator";
@@ -36,6 +37,7 @@ import SearchBar from "./components/searchBar";
 import { Skeleton, SkeletonLoader } from "./components/skeletonLoader";
 import CustomSlider from "./components/slider";
 import StatusBadge from "./components/statusBadge";
+import StructuredContent from "./components/structuredContent";
 import Tabs from "./components/tabs";
 import TemplateStyles from "./components/template";
 import Text from "./components/text";
@@ -94,6 +96,7 @@ const theme = extendTheme({
     Label,
     Link,
     List,
+    Logo,
     ...NotificationStyles,
     Pagination,
     ProgressIndicator,
@@ -104,6 +107,7 @@ const theme = extendTheme({
     SkeletonLoader,
     CustomSlider,
     StatusBadge,
+    StructuredContent,
     Tabs,
     ...TemplateStyles,
     Text,